3 done 	modified:   2/main.c 	new file:   3.1/codec.c 	new file:   3.1/codec.h 	new file:   3.1/decode.c 	new file:   3.1/encode.c 	new file:   3.1/input.txt 	new file:   3.1/makefile 	new file:   3.1/output.txt 	modified:   3/codec.c 	modified:   3/codec.h 	new file:   3/encode 	modified:   3/encode.c 	modified:   4/main 	modified:   4/myunzip.gpg 	new file:   4/myzip.gpg 	modified:   4/prog 	modified:   4/prog.c

+++ 2/main.c
@@ -1,9 +1,10 @@
 #include <stdio.h>
 #include <math.h>
+// need to add a margin of error epsilon for marginal errors?
 int main(){
     printf("enter 3 integers for triangle sides\n");
-    int a, b, c;
-    scanf("%d %d %d", &a, &b, &c);
+    double a, b, c;
+    scanf("%f %f %f", &a, &b, &c);
     if(pow(a,2) + pow(b,2) == pow(c,2)){
         double aa, ba, ca;
         //furmula found on calculator.net/triangle
